add SensorFaultHandler delegate
Each Sensor has an Action which is called when the value dips into danger
zone or satisfies a SensorTrigger.
SensorWatcher has a single Faulted delegate which is raised when any of
the Sensors Faulted Action is called.
ISensorAccess gets a SensorFaultHandler delegate so you can register a
function to be called whenever a Sensor enters faulted state.