filter FaultReports, enable active triggers
Filter the list of Fault Reports by MchCode (ObjectID in DeviceSettings).

Add a new list to Sensor of active triggers. These are updated when you
Save the Sensor, which will pull in all SensorTrigger from the sqlite db
and into a memory array or Sensor.Trigger which are the active triggers
that are compared against during each polling in the SensorWatcher thread.